BugFix: try and get both Mac and Linux Travis CI builds working
The Personal Package Archive (on Ubuntu) called `ppa:jonathonf/php7`
no longer seems to be available which has cause Linux CI builds to error
out however there is a commonly used (mainline ?) alternative called
`ppa:ondrej/php` which seems to do the job.

Also with recent upgrades the Mac CI builds are erroring because Boost 1.68
is being built from scratch as it is not available as a "bottled" formula
on the Homebrew system - and it takes so long the install/setup process
times out... However we do not need a bleeding edge Boost so it is
sufficient to use the supplied CI version (which seems to be 1.60).

Also fixed a spelling mistake in a comment in the CI/travis.osx.install.sh
file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>