Enhance: arrange for main menu tool tips to actually be shown to user
In https://github.com/Mudlet/Mudlet/pull/2110 I arranged for all the no-
longer shown status tips (we discarded the status bar on the main window
some time ago) to become tool tips for the main menu.  Unfortunately, I
never turned on the option for those tool-tips to be shown!

This commit fixes that.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>