BugFix: correct presentation of texts in 2D Mapper set room weight code
The use of QString::arg(...) with multiple (in these cases 2) numeric
arguments was triggering an overloaded version where the second argument
is a field-width rather than a second text item to display. This is fixed
by either using multiple chained `arg(...)` invocations with a single value
or forcing each number to be treated as a `QString` with the static
`QString::number(...)` conversion method.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>