BugFix: correct off-by-one indexing in SGR(3|4)8;5;#m
This was previously really messing SGR sequence handling up as the colour
index (the third number) in a SGR(3|4)8;5;#m sequence was being interpreted
as a SGR#m code which was most commonly showing up as odd lapses into using
italics and underlines when not expected.

I have also commented out some SGR#m `case #:`s in some `switch () {}`
that did not have any implementations so that they will be reported as
being un-handled rather than being silently dropped even when
`DEBUG_SGR_PROCESSING` is defined during debug builds.  This is so their
use can be spotted in the debug output in the future...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>