Add text analyser to consoles (#1579)
This will allow the user to determine the exact characters present in a
line of MUD (or locally generated) text which could be useful in working
out what to use to construct triggers and also how the displayed characters
are constructed in UTF-8 bytes.  To use it the user should control click
on a line of text in a console and then right click to bring up the context
menu.  By moving the cursor over the "analyser characters" action and
hovering there a generated tool-tip will be brought up that contains HTML
tables with the break down of Unicode code-points which is then broken down
into the corresponding UTF-16 (which Qt uses internally to store
all text in QStrings) and UTF-8 which is what the Lua sub-system gets as
its form of strings.  For characters that are not on the Basic Multi-Plane
and which thus need two QChar (and thus UTF-16 surrogate pairs) to
represent the pairs of UTF-16 values (in the range 0xD800 to 0xDBFF for the
first, high surrogate followed by one in the range 0xDC00 to 0xDFFF for the
second, low surrogate) those numbers are also displayed.

Note: this works on (only) the first line of the selected text even
if more than one line is selected.  The context menu entry is NOT shown if
there is no selection.

I found there are some non-printing Unicode code-points that it is sensible
to report - especially as they cannot currently be shown {the FitzPatrick
Emoji modifiers that give a skin colour tint to human related Emojis},
however as they are not on the BMP a revision was needed to:
`inline QString TTextEdit::convertWhiteSpaceToVisual(...)`

Added per profile option to enable (disabled by default) this analyser as
requested.

Added an additional line to the information table which gives the decimal
escape codes that would be needed to specify the individual bytes that
make up a non-ASCII grapheme in Lua for characters that are not printable
ASCII characters.

Previously this PR would show the entire first line that contained the
start of the selected text in the analysis - this commit revises things
so that only the selected part is shown. However the (1-based) indexes into
the QChar (QString, UTF16BE, C++ core view) and UTF-8 (Lua view) of the
selection are reworked so that they are still correct for a partially
selected line.

The connect(...) call has also been updated to the Qt5 style. Note that
this feature uses the QAction::hovered() signal which can cause problems
when used if running inside a debugger such as GDB from Qt Creator.

Some HTML/Rich text entries in the dlgProfilePreferences constructor got
reverted to a form with an unnecessary QStringLiteral wrapper containing
a commonly used set of outer tags.

Ran clang-format on the whole of ./src/TTextEdit.cpp

Revise: display full strings for non-visible characters in analyser

Although wanted by reviewer this might not be a good idea if large numbers
of Emoji characters are involved with FitzPatrick modifiers as the display
widget might need to be a bit large even though it is limited to showing
sixteen glyphs per row in the tabular layout - we will just have to see how it
works out.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>