Enhance: add lua getMapSelection() function {only covers rooms for now}
As an aside in: https://github.com/Mudlet/Mudlet/issues/2155
I suggested that we could use a lua function that would return a list of
rooms when they are selected in the (2D) mapper - this will be useful for
user scripts/aliases that want to modify such a selection.

To allow the handling when something other than a room or rooms is selected
the function actually returns an associative table which always contains
the key `type` which will be `rooms` if it does contain a selection of
rooms and `none` otherwise at the present. It should be possible to extend
this scheme later if desired.  If there is a `rooms` value for the `type`
then there will also be a `center` key with the room id number as the value
which will be the only selected room id or that of the room that has the
yellow cross-hairs in the 2D mapper; finally there will also be a key
`rooms` which will contain the (sorted in ascending order) list of
selected room ids.

As a zero room id is NOT a valid room number I have revised
`(int) T2DMap::mMultiSelectionHighlightRoomId` to have the `0` rather than
`-1` as the sentinel value for the *no selected room* case.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>