BugFix: allow XMLexport to files with non-Latin1 character filenames
In Issue https://github.com/Mudlet/Mudlet/issues/1872 it is reported that
non-ASCII characters in the file name for items that are to exported as
packages fails on Windows platforms.  It transpires that the MinGW STL
`std::stream` functions do not handle such characters as the Windows file
i/o sub-system needs non-ASCII filenames to be supplied as a wide character
array (using `wchar_t` type) but there is no requirement for that in the
C++ STL specification (it is a non-standard extension in MicroSoft's MSVC
library) and the third-party MinGW/MinGW64 STLs do not include it.

The workaround is to use `QFile` to provide the file handling and to push
the `pugixml` document through it as a `std::string` obtained from the
pugixml std::stringstream interface. We cannot use the pugixml file
handling which itself does contain means to get around this awkwardness
on the Windows platform because we also need to convert any ASCII control
characters that are not valid in XML 1.0 in the XML data into our own
custom replacements and the XML entities that pugixml uses are not
compatible with the ones that we had started to use before starting to
use the pugixml library and we do not use that library to *read* the XML
files (and it is not clear that it would even parse the replacement
entities it uses anyway).

Whilst debugging all of this it became clear that there was a second bug
in the code to replace the pugixml entities with our own replacements in
that pugixml uses 2 digit numbers in the XML entities even for ASCII
character codes less than 10 - but our search and replacement method was
looking for single digit numbers. I have fixed the method to look for
both.

Also:
`(void) XMLexport::replaceAll(std::string&, char, const std::string&)`
was not being used, instead only the method:
`(void) XMLexport::replaceAll(std::string&, const std::string,
 const std::string&)`
varient is active. I have thus remove the former from the application.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>