Revise: change <Host ...> save file attribute names per peer review
Ideally an XML file should be human readable as well as machine readable.
So the names of attributes and elements would, in the best case, make
sense on their own; we have not really gone along with that practice in the
past and instead have used the name of the internal variable which does
mean that some of the details in the XML files are not, perhaps as
comprehensible to those not intimately familiar with the C++ core of the
application. This being the case I hereby convert the names of the
attributes needed for the settings introduced as part of this PR from a
long but descriptive camel case text to the internal variable name - as I
was instructed to! 8-P

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>