Refactor&Enhance: reduce some spaghetti from TBuffer class... (#2133)
Refactor code to remove `(int) TBuffer::mCode[1024]` as fixed int array
- although name has been reused as a QString to replace
`(QString) TBuffer::code`.

Eliminate the use of several `goto DECODE'` and indeed, that label entirely
from `TBuffer::translateToPlainText(...)`.

Renamed some variables to better explain their meanings and class
relationship:
* (bool) TBuffer::gotESC ==> TBuffer::mGotESC
* (bool) TBuffer::gotHeader ==> TBuffer::mGotCSI

Some class members were being shadowed in some of the `TBuffer` methods as
arguments or were effectively duplicates:
* (int) TBuffer::fgColorR   ==#
* (int) TBuffer::mFgColorR  ==#==> TBuffer::mCurrentFgColorR
* (int) TBuffer::fgColorG   ==#
* (int) TBuffer::mFgColorG  ==#==> TBuffer::mCurrentFgColorG
* (int) TBuffer::fgColorB   ==#
* (int) TBuffer::mFgColorB  ==#==> TBuffer::mCurrentFgColorB
* (int) TBuffer::fgColorLightR ==> TBuffer::mCurrentFgColorLightR
* (int) TBuffer::fgColorLightG ==> TBuffer::mCurrentFgColorLightG
* (int) TBuffer::fgColorLightB ==> TBuffer::mCurrentFgColorLightB
* (int) TBuffer::bgColorR ==> TBuffer::mCurrentBgColorR
* (int) TBuffer::bgColorG ==> TBuffer::mCurrentBgColorG
* (int) TBuffer::bgColorB ==> TBuffer::mCurrentBgColorB

Needed to add a per profile configuration option (on "Special options" tab)
to handle cases where a SGR (3|4)8;2;...m code using only semicolon as
separator either includes or excludes a Colour Space Id as the third
(expected to be empty) parameter. This is not needed when colons are used
to sub-divide the parameter string.

Also added a per profile configuration option (on "Colour view" tab) so
that the user can control whether the MUD Server is allowed to use the
OSC P and R sequences to redefine the basic ANSI 16-colours as mentioned
above.

Refactoring of the code enabled the elimination of a number of state
booleans that were used to carry forward the state of SGR (3|4)8 codes
because the text that contains all SGR codes is now split out from the
byte-by-byte processing and decoded separately as a sub-string:
* (bool) TBuffer::mFgColorCode
* (bool) TBuffer::mBgColorCode
* (bool) TBuffer::mWaitingForHighColorCode
* (bool) TBuffer::mWaitingForMillionsColorCode
* (bool) TBuffer::mIsHighOrMillionsColorMode
* (bool) TBuffer::mIsHighOrMillionsColorModeForeground
* (bool) TBuffer::mIsHighOrMillionsColorModeBackground

Also the unused:
* (TMxpElement) TBuffer::mCurrentElement
* (std::string) TBuffer::tempLine
* (int) TBuffer::currentFgColorProperty

And the factored out:
* (int) TBuffer::codeRet
* (QCOlor) TBuffer::mFgColor
* (QCOlor) TBuffer::mBgColor

As part of deducing the behaviour of ANSI ESC sequences I realised that
the MXP decoding is considerably short of handling the full MXP
specifications as detailed at: https://www.zuggsoft.com/zmud/mxp.htm .

Also removed some unused members of other classes that I found whilst
investigating the `(int) TBuffer::mCode[1024]` member:
* (bool) Host::mCodeCompletion
* (bool) Host::mDisableAutoCompletion
* (int) Host::mMXPMode
* (int) TConsole::currentFgColorProperty

Whilst working on the profile preference dialogue I spotted that there was
a mis-nested layout (one layout inside another one) involving some of the
Discord options on the last tab near to where I added the option to handle
the optional inclusion of the Colour Space Identifier in 16M colour codes
which only uses the semi-colon as a parameter separator. So I have cleaned
up that at the same time...

Improved by add a compile time DEFINE to hide an MXP debug message to
cut down spammy qDebug() output when it isn't needed.

I have also commented out some SGR#m `case #:`s in some `switch () {}`
that did not have any implementations so that they will be reported as
being un-handled rather than being silently dropped even when
`DEBUG_SGR_PROCESSING` is defined during debug builds.  This is so their
use can be spotted in the debug output in the future...

Added code to ensure main console is updated after OSC ANSI colour redefinition

Ideally an XML file should be human readable as well as machine readable.
So the names of attributes and elements would, in the best case, make
sense on their own; we have not really gone along with that practice in the
past and instead have used the name of the internal variable which does
mean that some of the details in the XML files are not, perhaps as
comprehensible to those not intimately familiar with the C++ core of the
application. This being the case I have named the attributes needed for the
settings introduced as part of this PR after the internal variable name.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>