Enhance: add lua getMapSelection() function {only covers rooms for now} (#2160)
As an aside in: https://github.com/Mudlet/Mudlet/issues/2155
I suggested that we could use a lua function that would return a list of
rooms when they are selected in the (2D) mapper - this will be useful for
user scripts/aliases that want to modify such a selection.

The function returns an associative table which, if it does contain a
selection of rooms, there will be a `center` key with the room id
number as the value which will be the only selected room id or that
of the room that has the yellow cross-hairs in the 2D mapper; there
will also be a key `rooms` which will contain the (sorted in ascending
order) list of selected room ids. At present this only handles rooms
being selected in the 2D mapper and will return just an empty table
in other circumstances. It is likely that further selections will be added
specifically for map label(s) and a custom exit line - users should check
for the `rooms` table in the returned value to confirm it is a selection
of rooms.

As a zero room id is NOT a valid room number I have revised
`(int) T2DMap::mMultiSelectionHighlightRoomId` to have the `0` rather than
`-1` as the sentinel value for the *no selected room* case.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>