BugFix: get buttons working in menus on button toolbars
A single character defect meant that the `SIGNAL` from the push-buttons on
menus were not being connected to the SLOT that was expected to get the
signal in order to perform the required behaviour.

As such the `EAction::triggered(QAction*)` signal declaration is not
required and can be excised.

This will close https://github.com/Mudlet/Mudlet/issues/1914 .

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>