Refactor: remove all possible remaining old style SIGNAL/SLOT connect(...)s
This includes some unimplemented and commented out instances.

It does not include a usage of `QSignalMapper` which has been declared
obsolete and is not supported by the new, compile-time checking, system.

It also does not include useages in the third parties libraries,
particularly the edbee-lib editor widget and the communi IRC libraries.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>