Refactor: clean up deliberate mis-layout of MXP code in a TBuffer method
In https://github.com/Mudlet/Mudlet/pull/2133 I deliberately left some code
untouched so that the diffs were a bit more comprehensible - this was in
the region of the code that identified the CSI sequences associated with
setting the MXP mode.

This commit cleans up and lays out the code correctly to match the
surrounding indentation (and adds some related MXP mode comments), which
might be a preliminary for a solution to:
https://github.com/Mudlet/Mudlet/issues/1623

It should not change the functionality at all.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>