Refactor: move TConsole functions from mudlet class (Part 1)
As part of an attempt to rationalise the lua subsystem to Host API I want
to remove the need for many lua functions to have to call a method in the
main `mudlet` class if they are NOT directly interacting with the main
`TConsole` instance in a profile and calling a similar method in that
same instance if they do.  The current code maintained some `QMap`s in the
`mudlet` class that seem to duplicate the exact same types of container in
each of the "Main Console" instance for each profile.

This first PR removes the following `QMap`s from the `mudlet` class and
refactors the code to access the same data from each profile's main
`TConsole`:
- `(QMap<Host*, QMap<QString, TConsole*>>) mHostConsoleMap`
- `(QMap<Host*, QMap<QString, TDockWidget*>>) mHostDockConsoleMap`
- `(QMap<Host*, QMap<QString, TLabel*>>) mHostLabelMap`

Two of the equivalents already existed in the `TConsole` class but they
were `std::map` rather than `QMap` containers - however by converting them
to the Qt container type it was possible to use some more abbreviated code
using the handy convenience that `QMap<T1, T2>::value(T1)` automatically
returns a `nullptr` value if T2 is a pointer type and T1 is NOT present as
a key in the QMap.

Also it means that it is more obvious that some functions that currently
take a `std::string` argument can be refactored to take a `QString` type
in a future PR to uniformly support non-Latin1 item names very soon - at
present it needs further checking/fixing before I can be sure that all
buffer/mini consoles/user windows and/or labels can have non-Latin1 names!

As a side effect I was able to eliminate the following members/method
- `(bool) TTExtEdit::mIsDebugConsole`
- `(bool) TTExtEdit::mIsMiniConsole`
- `(void) TTExtEdit::setIsMiniConsole()`
as the TTextEdit class can determine the values the flags would have by
reference to the parent `TConsole` mType member.

Also removable were the unused:
- `(QList<TConsole*>) TConsole::mSubConsoleList`
- `(void) mudlet::slot_userToolBar_orientation_changed(Qt::Orientation)`
- `(void) mudlet::slot_userToolBar_hovered(QAction* pA)`

Finally this has enabled me to remove quite a few of the blatantly
confusing and frequent usage of the term **`dockWindowConsoleMap`** as a
`auto` scope reference to a `QMap` that contains all the mini consoles,
user windows and buffer instances of profile's user created `TConsole`s
when only the *user window* type ones **are dock-/float-able** -
compounding that is the use of another `auto` scope reference to a `QMap`
of `TDockWidget`s called **`dockWindowMap`** which DID refer to that
sub-set; I suspect some copy-pasting that should have been revised after
the paste but wasn't!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>