Enhance: extend getLines(...) to support user created windows
Allows a extra initial string argument as a mini-console, user window or
buffer name to get the text from `TConsole` instance as well as the main
profile one. To that end, the empty string and the name "main" are reserved
and treated as referring to the main one.

Note that this PR may clash with another PR I have pending:
https://github.com/Mudlet/Mudlet/pull/2226

This will close: https://github.com/Mudlet/Mudlet/issues/754 which was
originally opened in December 2013!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>