Enhance: allow Server to ECHO for us - to allow hiding of password entry
We now allow the Server to negotiate the Telnet ECHO (1) sub-option so that
he may provide the echoing of text we send to him.  The accepted practice
is that if he is doing that we will not do it ourself (to prevent doubled
lines) - and this is normally done during the sending of passwords so that
he does not actually echo anything (or perhaps sends back only asterisks)
so as to hide the entry on our screen.

This overrides the effect of our former "Echo the text you send in the
display box" setting. So that this is made obvious when it is happening
that control has been moved to a checkable toolbutton to the left of the
command line on the main TConsole for a profile.  Importantly it shows
a new blue icon I have made (from part of the Oxygen icon set icon:
"mail-reply-sender") which is show if the option is enabled and a grey
version to show if the control is disabled.  Whilst the Server is
performing any echoing however the control changes from the blue icon to
a "dialog-warning" yellow-triangular icon instead which is to alert the
user that what they type will NOT be shown on screen. This will not show
if the option is switched off because then there is no difference in
behaviour.

A suitable tooltip is provided in all instances so that the user can find
out what this feature is indicating in it's current state.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>