Revert: remove UI aspects of previous commit
Following peer-review it was intimated that having a UI element taking up
space in the command line area was unwanted.  This commit undoes the part
that added the combine control and indicator and reverts back to the
checkbox on the preferences dialogue. I have updated the tooltip for that
control to indicate that it's effects when enabled can be suppressed by the
game server.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>