Revise: reorganise cTelnet sub-option negotiation debug output
Some of the messages do not make sense for some options. Like the one that
says we do not want to enable one of OUR options when the Server is
actually informing us that his is disabling one of HIS options that were
previously active.

Removed the previously ambiguous #define `DEBUG` and merged it with the
`DEBUG_TELNET` - now it will turn on the display of WILL/WONT/DO/DONT
if defined and some extra details if defined as greater than 1.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>