Refactor: clean up TBuffer text format aspects (#1840)
This is the edited summary of a squash and merge of 14 commits:

Adds support for SGR Reverse (swap foreground and background colours)
"7"/"27" for On/Off and SGR Overline "53"/"55" for On/Off

Remove unused cruft:
* (void) TTextEdit::drawFrame(QPainter&, const QRect&)
* (void) TTextEdit::updateLastLine
* const QChar cLF & cSPACE in TBuffer (as it happens they are completely
  unused and redundant as the enum QChar::SpecialCharacter provides
  QChar::LineFeed and QChar::Space to provide the same constants)
* (QTime) TBuffer::mTime
* (void) TConsole::echoUserWindow(const QString&)
* (QPoint) TBuffer::insert(QPoint&, const QString&, int, int, int,
                           int, int, int, bool, bool, bool, bool)
* (void) TConsole::printDebug(...) functionally the same as one type of
  (void) TConsole::print(...) just with a different order of arguments.

Convert #define constants TCHAR_BOLD etc. into a QFLag/enum
TChar::AttributeFlags which is declared and capable of QFlag OR operations.

Refactor a number of methods that take lots of bools and ints as individual
formatting options and colour components to take single
TChar::AttributeFlags and one or two QColors instead.

Remove a large number of (int) colour value component values as member
variables in TBuffer as they are not needed.

Convert highly repetitive intermediate methods to setBold, setItalics etc.
to take a (combinations allowed) TChar::Attribute flag value instead.

Convert 2x3 int as colour components (r,g,b) in TColorTable defined in
TTrigger class to a pair of QColors.

Remove unused QString argument from:
* (void) mudlet::setLink(...)
* (void) TConsole::setLink(...)

Remove unused QColor argument from:
* (inline void) TTextEdit::drawCharacters(...)

Refactor arguments in:
*(QString) TBuffer::bufferToHtml(QPoint P1, QPoint P2,
                              bool allowedTimestamps, int spacePadding = 0)
to:
 (QString) TBuffer::bufferToHtml(const bool showTimeStamp = false,
                              const int row = -1, const int endColumn = -1,
                          const int startColumn = 0,  int spacePadding = 0)

Convert to const references some method arguments.

Add TBuffer::set[BF]gColor(...) overloads that take a QColor argument.
Add selection state methods select()/deselect()/isSelected() const methods
to TChar class to hide/separate the selection process from the formatting
effect. (TChar::Reverse tracks the ANSI SGR reverse colour attribute and
its effect is EX-ORed with the (bool) TChar::mIsSelected flag).

Add a new tempAnsiColorTrigger lua function that, unlike tempColorTrigger
uses the correct ANSIcolors in the range 0-15 - although the original also
handles the 256 colour range in the 16-255 correctly those first 16 values
are miss-mapped and it is not possible to change them without breakage.

Adds 256-color support to Editor GUI for color triggers - and allows
choosing the default (unmodified) fore or background colors to match one
(the previous did not) and also allows one of the fore or background color
to be ignored so only the other is considered. The ignored color case is
saved in the profile data and can exported but MAY not work in previous
Mudlet versions which cannot handle the value used! It is also reported as
an error to have a color trigger with both fore and background ignored in
both the lua functions and in the GUI.

Also:
Fixed a code structure issue in TLuaInterpreter::debug() which would not
work correctly if there was more than one value on the lua stack to print
out.

This will close issues #477 and #703.

Converted some `QObject::connect` calls to the new Qt5 compile time
version.

Removed an unused `TTrigger*` argument from
`dlgColorTrigger::setupBasicButtons(...)`.

Removed an unused flag:
* `(bool) TConsole::mSaveLayoutRequested

Also spotted some dead code from abandoned attempt to support blinking,
some reordering that a new version of Qt spotted as needed in the TBuffer
and TChar constructor initialisation lists, and a operator precedence item
that could be clarified with an addition pair of `(`...`)`s.

A previous error in the prior PR that this one is attempting to replace
had a problem in HTML generation that I reproduced here and which needed
the same fix (a missing escaped `"` mark).

An upgraded Qt Creator pointed out to me some initiliser list issues in
`TBuffer` and `TConsole`; and some C-style casts in some font settings in
the latter class.

Remove a debugging output line that is not useful now and will be spammy.

Modernise a triplet of QObject::connect(...) calls that will otherwise
clash harder when merged into development after "upgrade-to-qt5-connect"
PR has also been merged into the main branch.

Also:
* add some explanation/help text to the dlgColorTrigger dialog.
* revise the text explaining the formula (232 + grey scale value 0..23)
used for colours from the 24 grey-scale part of the 256 colour range.
* add some tool-tips to parts of the dlgColorTrigger dialog.
* add a generic static method to the mudlet class to provide a consistent
and re-usable "HTML" wrapper around text - which will be particularly
useful for tool-tip generation.

Revise: change colour trigger UI to hide colours 16-255 by default

This follows a suggestion from a peer in the review process.

Update: add Wiki documentation link comment

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>