Enhance: add functions for handling game character details
`getPlayerName` takes no arguments and returns the "Character name" from
the connection screen. There is no provision to change this stored setting.

`sendPlayerName` takes no arguments and causes the "Character name" from
the connection screen to be sent without echoing to the game server, it
will return a nil plus an error message should the name not be set (is
empty string).

`sendPlayerPassword` takes no arguments and causes the "Password" from
the connection screen to be sent without echoing to the game server, it
will return a nil plus an error message should no password be set (is empty
string). There is no provision to find out this stored setting.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>