Revise: clean up some TTimer issues
This will split up the `QMap` in the `mudlet` class that contains all the
`QTimer` & `TTimer` paired up in the `TTimer` class so that when a `Host`
instance runs its `resetProfile()` method only the QTimers that are
associated with the TTimers in THAT profile get removed from the `mudlet`
map.  Previously all of them were nuked in that table.

Cleaned up the `TTimer` class - renamed the `(QTimer*) mpTimer` to be
called `mpQTimer` because in some places we refer to instances of this
class with a `mpTimer` member variable and sometimes we want to access the
`QTimer` within and using the same name can lead to confusion.

This also lead me to discover that there was no need to include the latter
in the arguments to `mudlet::registerTimer(...)` as it is already availalbe
from the previous argument.

I also found out one cause of *bogus* "MUDLET CRITICAL ERROR: trying to
unregister Timer but it is not registered!" messages in that a TTimer is
created as a root item for `TTimer` items every time
`XMLimport::importPackage(...)` is run and it is delete at the end if the
package does not import any timer items. I have added a flag that can be
set in that method so that the (revised) error message is not produced in
*that* case.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>