BugFix: prevent update running on Mudlet exit when set to manual (LINUX)
Also stops it running on development version when it is supposed to be
permanently disabled but was still putting up the change-log when the main
application ended.

There is similar code in the Update::setupOnWndows() method and I do wonder
whether it may also be the cause of:
https://github.com/Mudlet/Mudlet/issues/1823
in the Windows version?

I also found two other things needed to keep the updater sedated for
development versions of the code:
* hide the Update option from the "About" sub-menu
* Prevent the setup of the UpdateDialog on start up as that initiates the
  on-line search for updates - which is pointless in the development
  version case.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>