Revise: remove handling of HUNSPELL_OK_WARN symbol
It turns out that it and HUNSPELL_OK were phased out between the 1.4.1 that
I have on my GNU/Linux OS and 1.5.0 which is earlier than our CI systems
use.  Given the fact that the bleeding edge version is around 1.7.x there
is little point in maintaining this feature.  It does seem that later
versions may have a different mechanism to report the same detail that this
obsolete symbols represent. The return value from spell(...) or rather the
C++ method Hunspell_spell(...) is defined to have a boolean return value
type.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>