BugFix: 2D mapper custom line creation
With the introduction of translations AND also after the conversion of the
internal exit direction keys for normal exits in the `TRoom` data
structures for custom exit lines, in:
https://github.com/Mudlet/Mudlet/pull/2106
(which was included in Mudlet 3.16.0) it has not possible to create new
custom exit lines without encountering a bug that means the lines are
drawn in black and do not show up whilst drawing, the properties dialogue
also has no idea of the destination room and there are other oddities.

This commit contains an extract from:
https://github.com/Mudlet/Mudlet/pull/1810
that was pending and had the fix for this problem

Note that this does not fix the lua getCustomLines(...) function which,
since 3.16.0 has also been using lower case keys to return the normal exit
direction custom exit line details.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>