BugFix: ensure command line spell checker handles non-Latin1 characters (#2269)
Using QString::toLatin1() when transferring text to and from the Hunspell
library will miss-encode any characters that are not in the ISO 8859-1
(Latin 1) character set and will not work if used with languages that use
characters outside of that set.

The existing code only installed the user selected dictionary for spell-
checking in the constructor of the `TCommandLine` class instances which
meant that changes with the profile active did not take effect until the
next profile load - which is less user friendly than changing it when
requested. This commit corrects that problem.

At the moment there is (or should be?) only one `TCommandLine` per profile
- for the Main Console.  I have chosen to use the signal/slot method for
the `Host` class to inform the rest of the profile that the dictionary
has been changed. This is so it can be extensible in the future if
https://github.com/Mudlet/Mudlet/issues/1897
is implemented, so that we could:
1. centralise access to a profile dictionary
2. provide a custom dictionary for each profile
3. give Lua API access to both a custom and indeed the selected dictionary

although we will probably want to consider moving the code accessing the
Hunspell library to the main profile `TConsole` class.

I have also added a qDebug() line to report on the loading of a dictionary,
during development of this PR the output from this showed me that each
dictionary could have a different character encoding and that it was
essential to use the indicated one when transferred text into and out of
the Hunspell library.

Remove useless cruft from class:
* (bool) TCommandLine::mAutoCompletion
* (QPalette) TCommandLine::mAutoCompletionPalette
* (QString) TCommandLine::mAutoCompletionTyped
* (QMap<QString, int>) TCommandLine::mHistoryMap
* (QString) TCommandLine::mLastCompletion
* (bool) TCommandLine::mTabCompletion
* (QPalette) TCommandLine::mTabCompletionPalette

Clean up initialisation list.

Improve the context menu action so that the spelling suggestions are
presented on top of the normal context menu rather than separately as
(before).

Fixup some CLazy else-after-return code style warnings.

The context menu spelling checker will now report "no suggestions" rather
than producing no information at all if it cannot find a suggestion which
helps to make it different from the case where the word is correctly
spelled.

Replace one use of `for (;;)` with the more obvious `forever`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>