Revert: go back to Sentence case context menu items
Also take out a double negative that is not that helpful in a tooltip and
convert a menu item from the plural to the singular case as it can only do
one item (custom exit line) per invocation.

This also introduces some non-breaking spaces and hyphens so that
fragments of the info texts are kept together should the display area be
too narrow to fit each line as a whole.

Due to at least one Qt bug it is not possible to put non-breaking spaces
within the source file directly and have to be introduced in a different
way - I have done it here by using the first reparable parameter (%1).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>