Enhance+BugFix: handle 100% complete translation and plurals-only en_US one
This commit allows the generate-translation-stat.lua file to run if there
is NOT a line about "Ignored 999 untranslated source text(s)" which will
not be present should the translation be 100% complete.

It also hides the `mudlet_en_US.ts` translation file from the `qmake`
TRANSLATIONS variable because otherwise it will automatically be processed
without a `-pluralsonly` option to `lupdate` and that will populate it with
hundreds of texts that are not plurals and do not need "translation".  This
does mean that that language will have to be handled manually (a Mudlet
Maker like myself can run it through Qt Linguist to provide the
single/plural texts as needed - which currently is 11 cases and which I
have done in this commit).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>