Enhance: add per profile dictionary capability
It is saved at the end of a session and the affix file (`profile.aff`) is
maintained to allow for the Hunspell suggestion capability to operate.
The word-list (`profile.dic`) file is checked at session start to
accommodate manual editing (adding/removing words) between sessions.

When a word typed on the command line is not found in the main (system
provided on non-Windows OSes) instead of the red (now wavy on Unix,
dotted on macOS) underline a cyan dashed underline is used to show that the
word was found in the profile's own dictionary. Words can also be added
and removed.

There are added options to populate the profile dictionary with all words
sent by the MUD Server that do not appear in the selected system
dictionary, with options to require a minimum character (actually
*grapheme*) count in a word before it is so added and also an option to
not add words containing digits. Other options may be considered later.

Also found out why the wavy / dashed underline was not showing in previous/
initial attempt.

Some lua commands are also added:
* `addWordToProfileDictionary((string) word)` returns true if the word was
added (and was not already in the dictionary)
* `removeWordFromProfileDictionary((string) word)` returns true if the word
was present and removed from the dictionary.
* `getProfileDictionaryWordList()` returns a (sorted into system locale
case insensitive order) list of ALL words in the profile dictionary.
* `spellCheckWord((string) word [,(bool)useProfileDictionary])` do a check
in the specified system dictionary (default) or if a second argument is
supplied and is `true` use the per profile one. Returns `true` if the word
is found present in the dictionary, and false if not.
* `spellSuggestWord((string) word [,(bool)useProfileDictionary])` does the
same sort of suggestions search as is done on the console commmand line and
returns a list of suggestions from the specified system dictionary
(default) or if a second argument is supplied and is `true` use the per
profile one.

**ALL** spell-checking activity is ceased if the option is disabled in the
preferences (this was not the case previously)!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>